Externalizing processors tab content from InstancePage
