+++ src/app/Instance/InstancePage/InstancePage.tsx
@@ -1,7 +1,6 @@
 import React, { useCallback, useEffect, useState } from "react";
 import {
   Alert,
-  Button,
   DescriptionList,
   DescriptionListDescription,
   DescriptionListGroup,
@@ -14,8 +13,6 @@ import {
   DropdownPosition,
   DropdownSeparator,
   DropdownToggle,
-  EmptyState,
-  EmptyStateIcon,
   PageSection,
   PageSectionVariants,
   Split,
@@ -29,32 +26,18 @@ import {
   Text,
   TextContent,
   TextVariants,
-  Title,
 } from "@patternfly/react-core";
-import { Link, useHistory, useLocation, useParams } from "react-router-dom";
+import { useHistory, useParams } from "react-router-dom";
 import { useTranslation } from "react-i18next";
 import { Breadcrumb } from "@app/components/Breadcrumb/Breadcrumb";
-import { CaretDownIcon, PlusCircleIcon } from "@patternfly/react-icons";
-import {
-  DEFAULT_PAGE_SIZE,
-  FIRST_PAGE,
-  TableWithPagination,
-} from "@app/components/TableWithPagination/TableWithPagination";
-import { IAction, IRow, IRowData } from "@patternfly/react-table";
-import { formatDistance } from "date-fns";
+import { CaretDownIcon } from "@patternfly/react-icons";
 import "./InstancePage.css";
 import { InstanceDetails } from "@app/Instance/InstanceDetails/InstanceDetails";
-import StatusLabel from "@app/components/StatusLabel/StatusLabel";
 import { useGetBridgeApi } from "../../../hooks/useBridgesApi/useGetBridgeApi";
 import PageHeaderSkeleton from "@app/components/PageHeaderSkeleton/PageHeaderSkeleton";
-import { TableWithPaginationSkeleton } from "@app/components/TableWithPaginationSkeleton/TableWithPaginationSkeleton";
-import { useGetProcessorsApi } from "../../../hooks/useProcessorsApi/useGetProcessorsApi";
-import { usePolling } from "../../../hooks/usePolling/usePolling";
 import { BridgeResponse } from "@rhoas/smart-events-management-sdk";
 import DeleteInstance from "@app/Instance/DeleteInstance/DeleteInstance";
-import { TableRow } from "@app/components/Table";
 import { canDeleteResource } from "@utils/resourceUtils";
-import DeleteProcessor from "@app/Processor/DeleteProcessor/DeleteProcessor";
 import {
   getErrorCode,
   isServiceApiError,
@@ -66,6 +49,7 @@ import { getErrorHandlingMethodByType } from "../../../types/ErrorHandlingMethod
 import ProcessorDetailConfigParameters from "@app/Processor/ProcessorDetail/ProcessorDetailConfigParameters";
 import { useGetSchemaApi } from "../../../hooks/useSchemasApi/useGetSchemaApi";
 import { ProcessorSchemaType } from "../../../types/Processor";
+import { ProcessorsTabContent } from "@app/Instance/InstancePage/ProcessorsTabContent";
 
 export interface InstanceRouteParams {
   instanceId: string;
@@ -74,16 +58,11 @@ export interface InstanceRouteParams {
 const InstancePage = (): JSX.Element => {
   const { instanceId } = useParams<InstanceRouteParams>();
   const { t } = useTranslation(["openbridgeTempDictionary"]);
-  const location = useLocation();
   const history = useHistory();
 
   const processorsTabRef = React.createRef<HTMLElement>();
   const errorHandlingTabRef = React.createRef<HTMLElement>();
 
-  const [currentPage, setCurrentPage] = useState<number>(FIRST_PAGE);
-  const [currentPageSize, setCurrentPageSize] =
-    useState<number>(DEFAULT_PAGE_SIZE);
-  const [totalRows, setTotalRows] = useState<number>();
   const [activeTabKey, setActiveTabKey] = useState<number | string>(0);
   const [isDropdownActionOpen, setIsDropdownActionOpen] =
     useState<boolean>(false);
@@ -104,20 +83,7 @@ const InstancePage = (): JSX.Element => {
     getBridge(instanceId);
   }, [getBridge, instanceId]);
 
-  const {
-    getProcessors,
-    processorListResponse,
-    isLoading: areProcessorsLoading,
-    error: processorsError,
-  } = useGetProcessorsApi();
-
   const { getSchema } = useGetSchemaApi();
-
-  const triggerGetProcessors = useCallback(
-    (): void =>
-      getProcessors(instanceId, currentPage, currentPageSize, undefined, true),
-    [currentPage, currentPageSize, getProcessors, instanceId]
-  );
 
   const getPageTitle = useCallback(
     (bridge?: BridgeResponse) => {
@@ -132,20 +98,6 @@ const InstancePage = (): JSX.Element => {
     },
     [t]
   );
-
-  usePolling(() => triggerGetProcessors(), 10000);
-
-  useEffect(
-    () => getProcessors(instanceId, FIRST_PAGE, DEFAULT_PAGE_SIZE),
-    [getProcessors, instanceId]
-  );
-
-  useEffect(() => {
-    if (processorListResponse) {
-      setCurrentPage(processorListResponse.page ?? FIRST_PAGE);
-      setTotalRows(processorListResponse.total ?? 0);
-    }
-  }, [processorListResponse]);
 
   useEffect(() => {
     if (bridgeError && axios.isAxiosError(bridgeError)) {
@@ -168,14 +120,7 @@ const InstancePage = (): JSX.Element => {
         );
       }
     }
-
-    if (processorsError) {
-      throw new ErrorWithDetail(
-        getPageTitle(bridge),
-        t("instance.errors.processorsListGenericError")
-      );
-    }
-  }, [bridge, bridgeError, getPageTitle, history, processorsError, t]);
+  }, [bridge, bridgeError, getPageTitle, history, t]);
 
   useEffect(() => {
     if (bridge?.error_handler?.type) {
@@ -207,73 +152,7 @@ const InstancePage = (): JSX.Element => {
     setActiveTabKey(eventKey);
   };
 
-  const processorsOverviewColumns = [
-    {
-      accessor: "name",
-      label: t("common.name"),
-      formatter: (value: IRowData, row?: IRow): JSX.Element => {
-        const processorId = (row as BridgeResponse)?.id ?? "";
-        return (
-          <Link
-            data-testid="tableProcessors-linkProcessor"
-            to={`${location.pathname}/processor/${processorId}`}
-          >
-            {value}
-          </Link>
-        );
-      },
-    },
-    {
-      accessor: "type",
-      label: t("common.type"),
-      formatter: (value: IRowData): string => {
-        const typeString = value as unknown as string;
-        return !typeString || !typeString.length
-          ? ""
-          : typeString.charAt(0).toUpperCase() +
-              typeString.slice(1).toLowerCase();
-      },
-    },
-    {
-      accessor: "submitted_at",
-      label: t("common.submittedAt"),
-      formatter: (value: IRowData): string => {
-        const date = new Date(value as unknown as string);
-        return formatDistance(date, new Date()) + " " + t("common.ago");
-      },
-    },
-    {
-      accessor: "status",
-      label: t("common.status"),
-      formatter: (value: IRowData): JSX.Element => {
-        const statusString = (value as unknown as string) ?? "";
-        return <StatusLabel status={statusString} />;
-      },
-    },
-  ];
-
-  const customToolbarElement = (
-    <Link to={`${location.pathname}/create-processor`}>
-      <Button ouiaId="create-processor" variant="primary">
-        {t("processor.createProcessor")}
-      </Button>
-    </Link>
-  );
-
-  const onPaginationChange = useCallback(
-    (pageNumber: number, pageSize: number): void => {
-      const correctPageNumber =
-        pageSize === currentPageSize ? pageNumber : FIRST_PAGE;
-      setCurrentPage(correctPageNumber);
-      setCurrentPageSize(pageSize);
-      getProcessors(instanceId, correctPageNumber, pageSize);
-    },
-    [currentPageSize, getProcessors, instanceId]
-  );
-
   const [showInstanceDeleteModal, setShowInstanceDeleteModal] = useState(false);
-  const [deleteProcessorId, setDeleteProcessorId] = useState("");
-  const [deleteProcessorName, setDeleteProcessorName] = useState("");
 
   const deleteInstance = (): void => {
     setShowInstanceDeleteModal(true);
@@ -284,40 +163,9 @@ const InstancePage = (): JSX.Element => {
     history.push(`/`);
   }, [history]);
 
-  const [showProcessorDeleteModal, setShowProcessorDeleteModal] =
-    useState(false);
-
-  const deleteProcessor = (id: string, name: string): void => {
-    setDeleteProcessorId(id);
-    setDeleteProcessorName(name);
-    setShowProcessorDeleteModal(true);
-  };
-
-  const handleOnDeleteProcessorSuccess = useCallback((): void => {
-    setShowProcessorDeleteModal(false);
-    getProcessors(instanceId, currentPage, currentPageSize);
-  }, [getProcessors, instanceId, currentPage, currentPageSize]);
-
-  const tableActions = (rowData: TableRow): IAction[] => [
-    {
-      title: t("common.delete"),
-      onClick: (): void => {
-        const id = (rowData.originalData as BridgeResponse).id;
-        const name = (rowData.originalData as BridgeResponse).name;
-        if (id && name) {
-          deleteProcessor(id, name);
-        }
-      },
-      isDisabled: !canDeleteResource(
-        (rowData.originalData as BridgeResponse).status
-      ),
-    },
-  ];
-
   return (
     <>
-      {(isBridgeLoading ||
-        (totalRows === undefined && areProcessorsLoading)) && (
+      {isBridgeLoading && (
         <>
           <PageHeaderSkeleton
             pageTitle={t("instance.loadingInstance")}
@@ -325,19 +173,9 @@ const InstancePage = (): JSX.Element => {
             hasLabel={false}
             totalTabs={2}
           />
-          <PageSection>
-            <TabContent id="instance-skeleton__page__tabs-processors">
-              <TableWithPaginationSkeleton
-                hasActionColumn={true}
-                columns={processorsOverviewColumns}
-                totalRows={currentPageSize}
-                customToolbarElement={customToolbarElement}
-              />
-            </TabContent>
-          </PageSection>
         </>
       )}
-      {bridge && processorListResponse?.items && (
+      {bridge && (
         <>
           <Drawer isExpanded={showInstanceDrawer}>
             <DrawerContent
@@ -444,123 +282,98 @@ const InstancePage = (): JSX.Element => {
                   />
                 </Tabs>
               </PageSection>
-              <PageSection>
-                <TabContent
-                  eventKey={0}
-                  id="instance-page__tabs-processors"
-                  ouiaId="processors"
-                  ref={processorsTabRef}
-                  aria-label="Processors tab"
-                >
-                  <TableWithPagination
-                    columns={processorsOverviewColumns}
-                    customToolbarElement={customToolbarElement}
-                    rows={processorListResponse.items}
-                    tableLabel={t(
-                      "openbridgeTempDictionary:processor.processorsListTable"
-                    )}
-                    isLoading={areProcessorsLoading}
-                    onPaginationChange={onPaginationChange}
-                    pageNumber={currentPage}
-                    pageSize={currentPageSize}
-                    totalRows={totalRows ?? 0}
-                    renderActions={({ row, ActionsColumn }): JSX.Element => (
-                      <ActionsColumn items={tableActions(row)} />
-                    )}
-                  >
-                    <EmptyState variant="large">
-                      <EmptyStateIcon icon={PlusCircleIcon} />
-                      <Title headingLevel="h4" size="lg">
-                        {t("processor.noProcessors")}
-                      </Title>
-                    </EmptyState>
-                  </TableWithPagination>
-                </TabContent>
-                <TabContent
-                  eventKey={1}
-                  id="instance-page__tabs-error-handling"
-                  ouiaId="error-handling"
-                  ref={errorHandlingTabRef}
-                  aria-label="Error handling tab"
-                  hidden
-                >
-                  <PageSection
-                    className="instance-page__tabs-error-handling__section"
-                    variant={PageSectionVariants.light}
-                    isFilled
-                  >
-                    <Stack hasGutter>
-                      <StackItem>
-                        <TextContent>
-                          <Text
-                            component={TextVariants.h2}
-                            ouiaId="error-handling-section"
-                          >
-                            {t("common.errorHandlingMethod")}
-                          </Text>
-                        </TextContent>
-                      </StackItem>
-                      <StackItem>
-                        <DescriptionList>
-                          <DescriptionListGroup key="error-handling-method">
-                            <DescriptionListTerm>
-                              {t("common.errorHandlingMethod")}
-                            </DescriptionListTerm>
-                            <DescriptionListDescription>
-                              {
-                                getErrorHandlingMethodByType(
-                                  bridge.error_handler?.type
-                                ).label
-                              }
-                            </DescriptionListDescription>
-                          </DescriptionListGroup>
-                          {schema && !schemaError && !schemaLoading && (
-                            <>
-                              <ProcessorDetailConfigParameters
-                                schema={schema}
-                                parameters={
-                                  bridge.error_handler?.parameters as {
-                                    [key: string]: unknown;
-                                  }
-                                }
-                              />
-                            </>
-                          )}
-                          {schemaError && (
-                            <Alert
-                              className="instance-page__tabs-error-handling__alert"
-                              ouiaId="error-schema"
-                              variant="danger"
-                              title={schemaError}
-                              aria-live="polite"
-                              isInline
-                            />
-                          )}
-                        </DescriptionList>
-                      </StackItem>
-                    </Stack>
-                  </PageSection>
-                </TabContent>
-              </PageSection>
             </DrawerContent>
           </Drawer>
-          <DeleteInstance
-            instanceId={bridge.id}
-            instanceName={bridge.name}
-            showDeleteModal={showInstanceDeleteModal}
-            onCanceled={(): void => setShowInstanceDeleteModal(false)}
-            onDeleted={handleOnDeleteInstanceSuccess}
-          />
-          <DeleteProcessor
-            bridgeId={instanceId}
-            processorId={deleteProcessorId}
-            processorName={deleteProcessorName}
-            showDeleteModal={showProcessorDeleteModal}
-            onCanceled={(): void => setShowProcessorDeleteModal(false)}
-            onDeleted={handleOnDeleteProcessorSuccess}
-          />
         </>
       )}
+      <PageSection>
+        <TabContent
+          eventKey={0}
+          id="instance-page__tabs-processors"
+          ouiaId="processors"
+          ref={processorsTabRef}
+          aria-label="Processors tab"
+        >
+          <ProcessorsTabContent
+            instanceId={instanceId}
+            pageTitle={getPageTitle(bridge)}
+          />
+        </TabContent>
+        {bridge && (
+          <TabContent
+            eventKey={1}
+            id="instance-page__tabs-error-handling"
+            ouiaId="error-handling"
+            ref={errorHandlingTabRef}
+            aria-label="Error handling tab"
+            hidden
+          >
+            <PageSection
+              className="instance-page__tabs-error-handling__section"
+              variant={PageSectionVariants.light}
+              isFilled
+            >
+              <Stack hasGutter>
+                <StackItem>
+                  <TextContent>
+                    <Text
+                      component={TextVariants.h2}
+                      ouiaId="error-handling-section"
+                    >
+                      {t("common.errorHandlingMethod")}
+                    </Text>
+                  </TextContent>
+                </StackItem>
+                <StackItem>
+                  <DescriptionList>
+                    <DescriptionListGroup key="error-handling-method">
+                      <DescriptionListTerm>
+                        {t("common.errorHandlingMethod")}
+                      </DescriptionListTerm>
+                      <DescriptionListDescription>
+                        {
+                          getErrorHandlingMethodByType(
+                            bridge.error_handler?.type
+                          ).label
+                        }
+                      </DescriptionListDescription>
+                    </DescriptionListGroup>
+                    {schema && !schemaError && !schemaLoading && (
+                      <>
+                        <ProcessorDetailConfigParameters
+                          schema={schema}
+                          parameters={
+                            bridge.error_handler?.parameters as {
+                              [key: string]: unknown;
+                            }
+                          }
+                        />
+                      </>
+                    )}
+                    {schemaError && (
+                      <Alert
+                        className="instance-page__tabs-error-handling__alert"
+                        ouiaId="error-schema"
+                        variant="danger"
+                        title={schemaError}
+                        aria-live="polite"
+                        isInline
+                      />
+                    )}
+                  </DescriptionList>
+                </StackItem>
+              </Stack>
+            </PageSection>
+          </TabContent>
+        )}
+      </PageSection>
+      <DeleteInstance
+        instanceId={bridge?.id}
+        instanceName={bridge?.name}
+        showDeleteModal={showInstanceDeleteModal}
+        onCanceled={(): void => setShowInstanceDeleteModal(false)}
+        onDeleted={handleOnDeleteInstanceSuccess}
+      />
     </>
   );
 };
